+++ main.py
@@ -8,57 +8,18 @@ from callback import StreamingLLMCallbackHandler
 from langchain_community.vectorstores import FAISS
 from langchain_community.document_loaders import WebBaseLoader
 from langchain.embeddings.huggingface import HuggingFaceEmbeddings
-from langchain.text_splitter import CharacterTextSplitter
+from langchain.text_splitter import CharacterTextSplitter,RecursiveCharacterTextSplitter
 from websockets.exceptions import ConnectionClosedOK
-from query_data import get_chain
+from chain import Retriever,RAGChain
+import uvicorn
+import argparse
 import json
+
 
 load_dotenv()
 openai_api_key = os.environ.get("OPENAI_API_KEY")
 templates = Jinja2Templates(directory="templates")
 app = FastAPI()
-
-urls =[
-        'https://python.langchain.com/docs/modules/data_connection/',
-        'https://python.langchain.com/docs/modules/data_connection/document_loaders/'
-    ]
-
-def read_urls(config_path: str = 'config.json'):
-    """
-    Read urls from the config file
-    """
-    with open(config_path) as f:
-        data = json.load(f)
-    return data['urls']
-
-def setup_retriever(path: str=None):
-    """
-    Function to setup the retriever using faiss,huggingface and webbase loader
-    """
-    ## 
-    embeddings = HuggingFaceEmbeddings()
-    if os.path.exists('url_db/'):
-        db = FAISS.load_local("url_db/", embeddings=embeddings)
-        return db
-    else:
-        loader = WebBaseLoader(urls)
-        documents = loader.load()
-        text_splitter = CharacterTextSplitter(chunk_size=1000, chunk_overlap=0)
-        docs = text_splitter.split_documents(documents)
-        # embeddings_model_name = "sentence-transformers/all-mpnet-base-v2"  
-        
-        db = FAISS.from_documents(docs, embeddings)
-        return db
-
-def search_retriever(query:str):
-    """
-    Function to search the retriever
-    """
-    search = db.similarity_search(query, k=2)
-    return search
-
-logging.info("Setting up retriever...")
-db = setup_retriever(path="url_db/")
 
 @app.get("/")
 async def index(request: Request):
@@ -72,12 +33,12 @@ async def test_chat(request: Request):
 
 @app.websocket("/ws")
 async def websocket_endpoint(websocket: WebSocket):
-
     logging.info("WebSocket listening...")
     await websocket.accept()
     logging.info("WebSocket accepted...")
     stream_handler = StreamingLLMCallbackHandler(websocket)
-    qa_chain = get_chain(stream_handler)
+    qa_chain = RAGChain(stream_handler=stream_handler).setup_chain()
+    #qa_chain = get_chain(stream_handler)
     while True:
         try:
             # Receive and send back the client message
@@ -89,7 +50,7 @@ async def websocket_endpoint(websocket: WebSocket):
             # Construct a response
             start_resp = ChatResponse(sender="bot", message="", type="start")
             await websocket.send_json(start_resp.dict())
-            search = search_retriever(user_msg)
+            search = retriever.search_retriever(user_msg)
             # Send the message to the chain and feed the response back to the client
             output = await qa_chain.ainvoke(
                 {
@@ -117,3 +78,15 @@ async def websocket_endpoint(websocket: WebSocket):
                 type="error",
             )
             await websocket.send_json(resp.dict())
+
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument('-p', '--proxy', help="Use if proxies are required", default=True)
+    args = parser.parse_args()
+    logging.info(args.proxy)
+    if args.proxy != "False":
+        logging.info("Using proxies...")
+    else:
+        logging.info("Not using proxies...")
+        uvicorn.run(app)
